refactor: standardize ExponentialBackoffWithJitter retry strategy
Extracted the `ExponentialBackoffWithJitter` and `RetryStrategy` from `output_parser.rs` into a new `retry.rs` module within the `src/agents/builtin/core` directory.

Updated `tool_executor_engine.rs` to use this unified backoff strategy instead of its ad-hoc implementation, fulfilling the codebase optimization directives to refactor unclear blocks and improve maintainability while ensuring consistency across LLM retry loops in the agent harness.
