--- src/agents/builtin/.test_ralph_progress.json
@@ -1,20 +1,0 @@
-{
-  "task_description": "test task",
-  "features": [
-    {
-      "name": "Step 1",
-      "status": "completed"
-    },
-    {
-      "name": "Step 2",
-      "status": "completed"
-    }
-  ],
-  "current_feature_index": 2,
-  "notes": [
-    "Initialized task and broken down into features.",
-    "Completed feature Step 1: default output",
-    "Completed feature Step 2: default output"
-  ],
-  "is_complete": false
-}
