+++ src/agents/builtin/retry.rs
@@ -1,0 +1,41 @@
+use std::time::Duration;
+
+pub trait RetryStrategy: Send + Sync {
+    fn next_backoff(&self, attempt: usize) -> Duration;
+}
+
+pub struct ExponentialBackoffWithJitter {
+    base_ms: u64,
+    jitter_max_ms: u64,
+}
+
+impl Default for ExponentialBackoffWithJitter {
+    fn default() -> Self {
+        Self {
+            base_ms: 500,
+            jitter_max_ms: 100,
+        }
+    }
+}
+
+impl ExponentialBackoffWithJitter {
+    pub fn new(base_ms: u64, jitter_max_ms: u64) -> Self {
+        Self {
+            base_ms,
+            jitter_max_ms,
+        }
+    }
+}
+
+impl RetryStrategy for ExponentialBackoffWithJitter {
+    fn next_backoff(&self, attempt: usize) -> Duration {
+        let base_backoff = self.base_ms * (1 << attempt);
+        use rand::Rng;
+        let jitter = if self.jitter_max_ms > 0 {
+            rand::thread_rng().gen_range(0..self.jitter_max_ms)
+        } else {
+            0
+        };
+        Duration::from_millis(base_backoff + jitter)
+    }
+}
